Update README.md
This project focuses on building a machine learning model to predict whether a customer will make a purchase based on various personal and behavioral attributes. The dataset used for this analysis is a synthetic Purchase Info Dataset that contains features such as CustomerID, Age, Gender, PurchaseAmount, ProductCategory, and the target variable Purchased, which indicates whether the customer completed a purchase ("Yes") or not ("No").